third_party/WebKit/LayoutTests/media/track/regions-webvtt/text-track-cue-region-attribute.html

<!DOCTYPE html>
<title>Tests the regionId attribute of a cue.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var cue = new VTTCue(0, 1, 'sample');

    assert_equals(cue.regionId, '');
    cue.regionId = 'someId';
    assert_equals(cue.regionId, 'someId');

    var video = document.createElement('video');
    var track = document.createElement('track');
    track.onload = t.step_func_done(function() {
        cue = track.track.cues[0];
        assert_equals(cue.regionId, 'someregionattributeid');

        for (i = 1; i < 4; ++i) {
          cue = track.track.cues[i];
          assert_equals(cue.regionId, '');
        }
    });
    track.src = '../captions-webvtt/header-regions.vtt';
    track.kind = 'captions';
    track.default = true;
    video.appendChild(track);
});
</script>
